Remove upper-bound from skylight.
This is actually a (hopefully) temporary solution to a different problem. Although skylight levels can be easily scaled up in the same render method that affects blocklight, they cannot be scaled down like this. There is an override for the upper-bound *somewhere* else in the source, and I can't find it. The upper-bound slider was removed so anyone using the mod doesn't get confused as to why they can't make it dark during the day.